requirements.txt
pyswip
RPi.GPIO
opencv-python-headless